requirements.txt
Django==2.2.3
django-filter==2.1.0
pytz==2019.1
sqlparse==0.3.0
